% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coefChange}
\alias{coefChange}
\alias{compareModels}
\title{Compare two logistic/probit regression coefficient using different methods.}
\usage{
coefChange(reduced, full, method = "ystand", ratio.type = "perc", R = 1)

compareModels(reduced, full, method = "ystand", ratio.type = "perc")
}
\arguments{
\item{R}{Integer. If larger than \code{1}, confidence interval of the change are computed using bootstrap.}

\item{reduced}{The reduced model.}

\item{full}{The full model}

\item{method}{Character. The method used to compare coefficients (see
details).}

\item{ratio.type}{Character specifying how to compare the coefficients. This
should be one of \code{"perc"} (percentage change between the two models),
\code{"ratio"} (coefficient of the reduced model divided the coefficient of the full
one) or \code{"diff"} (difference between the two models).}
}
\description{
Compare two logistic/probit regression coefficient using different methods.
}
\details{
The \code{compareModels} function compares the coefficients of two logistic/probit
regressions using different methods. \itemize{
\item If \code{method="naive"}, the coefficients are compared without any transformation. This method is mainly
provided to highlight the effect of other methods.
\item If \code{method="ystand"}, y-standardization is used to compare the coefficients.
\item If \code{method="ape"}, Coefficients change are computed using average partial effects.
\item Finally, if \code{method="khb"}, the KHB method is used to compare the coefficients accross models.
}

The \code{coefChange} function only compute the change of coefficients using either
the percentage of change or the ratio of change. It can also compute
confidence intervals of this change using bootstrapping.
}

