% Generated by roxygen2 (4.0.0): do not edit by hand
\name{khb}
\alias{khb}
\alias{print.khb}
\title{KHB}
\usage{
khb(reduced, full, corrected.se = FALSE, med.sandwich = NULL,
  glm.sandwich = FALSE, savefs = FALSE)

print.khb(x, type = "summary", keyvar = NULL, disentangle = FALSE, ...)
}
\arguments{
\item{reduced}{The reduced model to compare (withought mediators variables). This
must be a logistic or probit regression model fit using the \code{glm}
command.}

\item{full}{The full model to compare (including mediators variables). This
must be a logistic or probit regression model fit using the \code{glm}
command.}

\item{corrected.se}{Logical. If \code{TRUE}, standard errors of the adjusted model
are corrected to account for the uncertainty of the regressions on the mediators
variables.}

\item{med.sandwich}{Logical. If \code{TRUE}, sandwich standard errors are
used for the regressions relating mediators and key variables. If
\code{NULL} (default), sandwich standard errors are used only if at least
one of the mediator is a binary (or categorical) variable.}

\item{glm.sandwich}{Logical. If \code{TRUE}, sandwich standard errors are
used for the glm models.}

\item{savefs}{Logical. If \code{TRUE}, the SUR model is saved in the result.}

\item{x}{A khb object}

\item{type}{Character. The type of information printed (see details).}

\item{keyvar}{A character vector or \code{NULL} (default). The list of
variables of interest (for which the change between models are displayed).
If \code{NULL} (default), all variables of the reduced models are considered as key
variables.}

\item{disentangle}{Logical. If \code{FALSE} (default), the contribution of each mediator is not printed.}

\item{...}{Arguments passed to other methods.}
}
\value{
a khb object which is a list with the following elements:\itemize{
   \item sum_conf a summary table of the confounding and rescaling effects.
   \item key a list of the confounding effect on each key variables (i.e. variable in the reduced model).
   \item reduced the reduced model.
   \item adjusted the adjusted model (with mediators residuals).
   \item full the full model.
   \item keyvar A character vector of the key variables (i.e. variable in the reduced model).
   \item mediators A character vector of the mediators (i.e. the variable in the full model
   that are not in the reduced one).
   }
}
\description{
Compare two non-linear regression models using the KHB method
}
\examples{
## Example: transition to higher education in Northern Ireland
data(mvad, package="TraMineR")
## Build a binary covariate, TRUE if went to higher education
HE <- rowSums(mvad[, 17:86]=="HE")>0

## Our reduced models including Father unemployement (funemp)
## and type of compulsory school (Grammar)
red <- glm(HE~funemp+Grammar, data=mvad, family=binomial("logit"))
## In the full model, we add the results at the end of compulsory schooling
## (gcse5eq).
full <- glm(HE~Grammar+funemp+gcse5eq, data=mvad, family=binomial("logit"))
## Running the KHB method
k <- khb(red, full)

## Printing the results for the funemp covariates only
print(k, keyvar="funemp")
## Comparing the models
print(k, type="models")

## Several mediators
## Here, we only have one covariate in the reduced model
red <- glm(HE~funemp, data=mvad, family=binomial("logit"))
full <- glm(HE~Grammar+gcse5eq+funemp, data=mvad, family=binomial("logit"))
#' ## Running the KHB method
k <- khb(red, full)

## Printing the results and disentangle the effect of each mediator
print(k, disentangle=TRUE)
## Comparing the models
print(k, type="models")
}
\references{
Karlson KB, Holm A and Breen R (2012). Comparing Regression Coefficients
  Between Same-sample Nested Models Using Logit and Probit: A New Method.
  Sociological Methodology, 42(1), pp 286-313.
}

