% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ystandcoef}
\alias{apecoef}
\alias{ystandcoef}
\title{Compute comparable coefficients accross models.}
\usage{
ystandcoef(x)

apecoef(x)
}
\arguments{
\item{x}{A logistic or probit regression.}
}
\value{
The y-standardized coefficients in which case the \code{"sdy"} attribute of the returned object store the "y" standard deviation or the average partial effects of each covariates.
}
\description{
The \code{ystandcoef} returns the y-standardized coefficients. This is done by dividing each coefficients by the standard deviation of the latent y dependant variable.
}
\details{
The \code{apecoef} returns the average partial effect of each coefficients.
}

